feat: add command to copy AWS Console URL
