+++ package.json
@@ -34,6 +34,10 @@
         "title": "Open AWS Console"
       },
       {
+        "command": "amplify-backend-explorer.copyUrl",
+        "title": "Copy AWS Console URL to clipboard"
+      },
+      {
         "command": "amplify-backend-explorer.switchCredentials",
         "title": "Switch AWS Profile"
       },
@@ -82,6 +86,10 @@
       "view/item/context": [
         {
           "command": "amplify-backend-explorer.openConsole",
+          "when": "view == amplify-backend-explorer && viewItem == resourceNode"
+        },
+        {
+          "command": "amplify-backend-explorer.copyUrl",
           "when": "view == amplify-backend-explorer && viewItem == resourceNode"
         },
         {
